+++ package.json
@@ -30,7 +30,6 @@
     "@apollo/client": "3.4.16",
     "@artsy/fresnel": "1.9.0",
     "@auth0/nextjs-auth0": "1.6.1",
-    "@emotion/babel-plugin": "11.3.0",
     "@emotion/is-prop-valid": "0.8.8",
     "@emotion/react": "11.4.1",
     "@emotion/styled": "11.3.0",
@@ -66,7 +65,7 @@
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-hook-form": "6.15.8",
-    "react-leaflet": "2.8.0",
+    "react-leaflet": "3.2.2",
     "react-leaflet-control": "2.1.2",
     "react-tooltip": "4.2.21",
     "resize-observer-polyfill": "1.5.1",
@@ -75,6 +74,7 @@
   "devDependencies": {
     "@commitlint/cli": "13.2.1",
     "@commitlint/config-conventional": "13.2.0",
+    "@emotion/babel-plugin": "11.3.0",
     "@graphql-codegen/cli": "2.2.1",
     "@graphql-codegen/import-types-preset": "2.1.6",
     "@graphql-codegen/typescript": "2.2.4",
